pyproject.toml
[project]
name = "tracking-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "ahnsv", email = "ahnsv@bc.edu" }
]
requires-python = ">=3.13"
dependencies = [
    "code-splitter>=0.1.5",
    "drtail-prompt",
    "gitpython>=3.1.44",
    "google-adk>=0.5.0",
    "google-cloud-aiplatform>=1.93.0",
    "google-cloud-storage>=2.19.0",
    "google-genai>=1.15.0",
    "llama-index>=0.12.36",
    "structlog>=25.3.0",
]

[project.scripts]
tracking-agent = "tracking_agent:main"
rag-ingest = "tracking_agent.ingestion.rag_corpus:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
drtail-prompt = { git = "https://github.com/drtail/drtail-prompt.git", rev = "v0.3.0" }
